roll stuff into source-info
