+++ vsc-extension/src-shared/source-info.test.ts
@@ -1,5 +1,4 @@
-import { findMatchingFunction } from "./serialized-source";
-import { MaybeSelectedFunction } from "./source-info";
+import { findMatchingFunction, MaybeSelectedFunction } from "./source-info";
 
 describe("findMatchingFunction", () => {
   it("should return undefined if no selected function is provided", () => {
