+++ vsc-extension/src-shared/source-info.ts
@@ -78,3 +78,22 @@ export type SelectedFunctionTestCases = Record<string, MaybeSelectedTestCase>;
 
 /** Map of filename -> functionName -> selected test index */
 export type SelectedFileTestCases = Record<string, SelectedFunctionTestCases>;
+
+export function findMatchingFunction(
+  sources: SerializableSourceFileMap,
+  selectedFunction: MaybeSelectedFunction
+) {
+  if (!selectedFunction) {
+    return undefined;
+  }
+  const matchingSource = Object.values(sources).find(
+    (source) => source.sourceFile.fileName === selectedFunction.fileName
+  );
+  if (!matchingSource) {
+    return undefined;
+  }
+  const matchingFunction = matchingSource.functions.find(
+    (fn) => fn.name === selectedFunction.functionName
+  );
+  return matchingFunction;
+}
